Dhruv_Rathod-binance-bot/requirements.txt
python-dotenv
python-binance
requests
